+++ test/integration/test_b2_command_line.py
@@ -223,18 +223,19 @@ def print_output(status, stdout, stderr):
 
 class Api:
     def __init__(
-        self, account_id, application_key, general_bucket_name_prefix, this_run_bucket_name_prefix
+        self, account_id, application_key, realm, general_bucket_name_prefix,
+        this_run_bucket_name_prefix
     ):
         self.account_id = account_id
         self.application_key = application_key
-
+        self.realm = realm
         self.general_bucket_name_prefix = general_bucket_name_prefix
         self.this_run_bucket_name_prefix = this_run_bucket_name_prefix
 
         info = InMemoryAccountInfo()
         cache = InMemoryCache()
         self.api = B2Api(info, cache=cache)
-        self.api.authorize_account('production', self.account_id, self.application_key)
+        self.api.authorize_account(self.realm, self.account_id, self.application_key)
 
     def create_bucket(self):
         bucket_name = self.this_run_bucket_name_prefix + '-' + random_hex(24)
@@ -314,10 +315,11 @@ class CommandLine:
         ),
     ]
 
-    def __init__(self, command, account_id, application_key, bucket_name_prefix):
+    def __init__(self, command, account_id, application_key, realm, bucket_name_prefix):
         self.command = command
         self.account_id = account_id
         self.application_key = application_key
+        self.realm = realm
         self.bucket_name_prefix = bucket_name_prefix
 
     def run_command(self, args, additional_env: Optional[dict] = None):
@@ -380,7 +382,12 @@ class CommandLine:
     def reauthorize(self):
         """Clear and authorize again to the account."""
         self.should_succeed(['clear-account'])
-        self.should_succeed(['authorize-account', self.account_id, self.application_key])
+        self.should_succeed(
+            [
+                'authorize-account', '--environment', self.realm, self.account_id,
+                self.application_key
+            ]
+        )
 
     def list_file_versions(self, bucket_name):
         return self.should_succeed_json(['ls', '--json', '--recursive', '--versions', bucket_name])
@@ -573,7 +580,9 @@ def key_restrictions_test(b2_tool, bucket_name):
     )
     key_one_id, key_one = created_key_stdout.split()
 
-    b2_tool.should_succeed(['authorize-account', key_one_id, key_one],)
+    b2_tool.should_succeed(
+        ['authorize-account', '--environment', b2_tool.realm, key_one_id, key_one],
+    )
 
     b2_tool.should_succeed(['get-bucket', bucket_name],)
     b2_tool.should_succeed(['get-bucket', second_bucket_name],)
@@ -590,7 +599,9 @@ def key_restrictions_test(b2_tool, bucket_name):
     )
     key_two_id, key_two = created_key_two_stdout.split()
 
-    b2_tool.should_succeed(['authorize-account', key_two_id, key_two],)
+    b2_tool.should_succeed(
+        ['authorize-account', '--environment', b2_tool.realm, key_two_id, key_two],
+    )
     b2_tool.should_succeed(['get-bucket', bucket_name],)
     b2_tool.should_succeed(['ls', bucket_name],)
 
@@ -601,7 +612,12 @@ def key_restrictions_test(b2_tool, bucket_name):
     b2_tool.should_fail(['ls', second_bucket_name], failed_list_files_err)
 
     # reauthorize with more capabilities for clean up
-    b2_tool.should_succeed(['authorize-account', b2_tool.account_id, b2_tool.application_key])
+    b2_tool.should_succeed(
+        [
+            'authorize-account', '--environment', b2_tool.realm, b2_tool.account_id,
+            b2_tool.application_key
+        ]
+    )
     b2_tool.should_succeed(['delete-bucket', second_bucket_name])
     b2_tool.should_succeed(['delete-key', key_one_id])
     b2_tool.should_succeed(['delete-key', key_two_id])
@@ -621,8 +637,16 @@ def account_test(b2_tool, bucket_name):
     bad_application_key = random_hex(len(b2_tool.application_key))
     b2_tool.should_fail(
         ['authorize-account', b2_tool.account_id, bad_application_key], r'unauthorized'
+    )  # this call doesn't use --environment on purpose, so that we check that it is non-mandatory
+    b2_tool.should_succeed(
+        [
+            'authorize-account',
+            '--environment',
+            b2_tool.realm,
+            b2_tool.account_id,
+            b2_tool.application_key,
+        ]
     )
-    b2_tool.should_succeed(['authorize-account', b2_tool.account_id, b2_tool.application_key])
     tearDown_envvar_test('B2_ACCOUNT_INFO')
 
     # Testing (B2_APPLICATION_KEY, B2_APPLICATION_KEY_ID) for commands other than authorize-account
@@ -2038,7 +2062,7 @@ def _assert_file_lock_configuration(
         assert legal_hold == actual_legal_hold
 
 
-def main(general_bucket_name_prefix, this_run_bucket_name_prefix):
+def main(realm, general_bucket_name_prefix, this_run_bucket_name_prefix):
     test_map = {  # yapf: disable
         'account': account_test,
         'basic': basic_test,
@@ -2072,9 +2096,11 @@ def main(general_bucket_name_prefix, this_run_bucket_name_prefix):
     if os.environ.get('B2_ACCOUNT_INFO') is not None:
         del os.environ['B2_ACCOUNT_INFO']
 
-    b2_tool = CommandLine(args.command, account_id, application_key, this_run_bucket_name_prefix)
+    b2_tool = CommandLine(
+        args.command, account_id, application_key, realm, this_run_bucket_name_prefix
+    )
     b2_api = Api(
-        account_id, application_key, general_bucket_name_prefix, this_run_bucket_name_prefix
+        account_id, application_key, realm, general_bucket_name_prefix, this_run_bucket_name_prefix
     )
 
     # Run each of the tests in its own empty bucket
@@ -2109,7 +2135,8 @@ def main(general_bucket_name_prefix, this_run_bucket_name_prefix):
 
 
 def cleanup_hook(
-    application_key_id, application_key, general_bucket_name_prefix, this_run_bucket_name_prefix
+    application_key_id, application_key, realm, general_bucket_name_prefix,
+    this_run_bucket_name_prefix
 ):
     print()
     print('#')
@@ -2117,7 +2144,8 @@ def cleanup_hook(
     print('#')
     print()
     b2_api = Api(
-        application_key_id, application_key, general_bucket_name_prefix, this_run_bucket_name_prefix
+        application_key_id, application_key, realm, general_bucket_name_prefix,
+        this_run_bucket_name_prefix
     )
     b2_api.clean_buckets()
 
@@ -2134,20 +2162,24 @@ def test_integration(sut, cleanup):
 
     print()
 
+    realm = os.environ.get('B2_TEST_ENVIRONMENT', 'production')
+
     sys.argv = ['test_b2_command_line.py', '--command', sut]
     general_bucket_name_prefix = 'test-b2-cli-'
     this_run_bucket_name_prefix = general_bucket_name_prefix + random_hex(8)
 
-    if cleanup:
-        atexit.register(
-            cleanup_hook, application_key_id, application_key, general_bucket_name_prefix,
-            this_run_bucket_name_prefix
-        )
-
-    main(general_bucket_name_prefix, this_run_bucket_name_prefix)
+    try:
+        main(realm, general_bucket_name_prefix, this_run_bucket_name_prefix)
+    finally:
+        if cleanup:
+            cleanup_hook(
+                application_key_id, application_key, realm, general_bucket_name_prefix,
+                this_run_bucket_name_prefix
+            )
 
 
 if __name__ == '__main__':
     general_bucket_name_prefix = 'test-b2-cli-'
+    realm = os.environ.get('B2_TEST_ENVIRONMENT', 'production')
     this_run_bucket_name_prefix = general_bucket_name_prefix + random_hex(8)
-    main(general_bucket_name_prefix, this_run_bucket_name_prefix)
+    main(realm, general_bucket_name_prefix, this_run_bucket_name_prefix)