+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add possibility to change realm during integration tests
+
 ## [2.5.0] - 2021-05-22
 
 ### Added
@@ -190,6 +193,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Add support for CORS rules in `create-bucket` and `update-bucket`.  `get-bucket` will display CORS rules.
+
+### Fixed
+* cleanup in integration tests works
 
 ## [1.0.0] - 2017-11-09